DATAOPS-716 Use customized list of adapters, FastQC (#60)
* Use customized list of adapters, FastQC

When hoovering over the adapter content plot in the the FastQC output
displayed in the end of all seqreports the name of the adapter found
is appearing. The adapters and names used are by default the ones
specified in the FastQC software. However, the provided list is outdated
and many of the adapters are used in several library preparation kits. A
customized list can now be used as input instead to look for adapters we
use at SNP&SEQ and to display relevant names when found.

Files changed:
- New file, adapter_list_fastqc.txt, added in config/tool_config/ to
  specify the adapters to look for.
- The fastqc command in main.nf is now updated with the additional
  a-flag that is used to specify the use of the file described under the
previous point.

* Update config/tool_config/adapter_list_fastqc.txt

Co-authored-by: Matilda Åslin <matilda.aslin@medsci.uu.se>

* Update config/tool_config/adapter_list_fastqc.txt

Co-authored-by: Matilda Åslin <matilda.aslin@medsci.uu.se>

* Fixed indentation and permalink

---------

Co-authored-by: Matilda Åslin <matilda.aslin@medsci.uu.se>